저장 완료, 추후에 몬스터와 데이터 수정 예정

+++ Train/Assets/Data_Game/Info_Player.cs
@@ -6,7 +6,7 @@ public class Info_Player
     public string Player_Name;
     public int Player_Atk;
     public int Player_Armor;
-    public int Player_Delay;
+    public float Player_Delay;
     public int Player_HP;
     public int Player_MoveSpeed;
     public string Player_Gun;
